+++ src/internal/utils.ts
@@ -1,4 +1,4 @@
-import { cloneDeepWith, mergeWith } from 'lodash';
+import { mergeWith, omit } from 'lodash';
 import * as mongoose from 'mongoose';
 import { format } from 'util';
 
@@ -15,7 +15,7 @@ import {
   WhatIsIt
 } from '../types';
 import { DecoratorKeys } from './constants';
-import { constructors, schemas } from './data';
+import { constructors, globalOptions, schemas } from './data';
 import { NoValidClass } from './errors';
 
 /**
@@ -222,10 +222,10 @@ export function mergeMetadata<T = any>(key: DecoratorKeys, value: unknown, cl: n
   }
 
   // Please dont remove the other values from the function, even when unused - it is made to be clear what is what
-  const current = cloneDeepWith(Reflect.getMetadata(key, cl) ?? {}, (val, ckey, obj, stack) => customMerger(key, val));
-
-  return mergeWith({}, current, value,
-    (objValue, srcValue, ckey, object, source, stack) => customMerger(key, srcValue));
+  return mergeWith({},
+    Reflect.getMetadata(key, cl),
+    value,
+    (objValue, srcValue, ckey, object, source, stack) => customMerger(ckey, srcValue));
 }
 
 /**
@@ -259,18 +259,21 @@ export function mergeSchemaOptions<T, U extends AnyParamConstructor<T>>(value: m
  * @param cl The Class
  */
 export function getName<T, U extends AnyParamConstructor<T>>(cl: U) {
-  const options: IModelOptions = Reflect.getMetadata(DecoratorKeys.ModelOptions, cl) ?? {};
-  const baseName = cl.name;
+  const options: IModelOptions =
+    Reflect.getMetadata(DecoratorKeys.ModelOptions, cl) ??
+    Reflect.getMetadata(DecoratorKeys.ModelOptions, cl.constructor) ??
+    {};
+  const baseName = cl.name ?? cl.constructor.name;
 
-  if (options.options && options.options.automaticName) {
-    const suffix = options?.options?.customName ?? options?.schemaOptions?.collection;
+  if (options.options?.automaticName) {
+    const suffix = options.options?.customName ?? options.schemaOptions?.collection;
 
     return !isNullOrUndefined(suffix) ? `${baseName}_${suffix}` : baseName;
   }
 
-  if (typeof options?.options?.customName === 'string') {
+  if (typeof options.options?.customName === 'string') {
     if (options.options.customName.length <= 0) {
-      throw new TypeError(`"customName" must be a string AND at least one character ("${cl.name}")`);
+      throw new TypeError(`"customName" must be a string AND at least one character ("${baseName}")`);
     }
 
     return options.options.customName;
@@ -288,7 +291,7 @@ export function isNotDefined(cl: any) {
     !isPrimitive(cl) &&
     cl !== Object &&
     cl !== mongoose.Schema.Types.Buffer &&
-    isNullOrUndefined(schemas.get(getName(cl)));
+    !schemas.has(getName(cl));
 }
 
 /**
@@ -405,4 +408,15 @@ export function warnMixed(target: any, key: string | symbol): void | never {
  */
 export function isNullOrUndefined(val: unknown): val is null | undefined {
   return val === null || val === undefined;
+}
+
+/**
+ * Assign Global ModelOptions if not already existing
+ * @param target Target Class
+ */
+export function assignGlobalModelOptions(target: any) {
+  if (isNullOrUndefined(Reflect.getMetadata(DecoratorKeys.ModelOptions, target))) {
+    logger.info('Assigning global Schema Options to "%s"', getName(target));
+    assignMetadata(DecoratorKeys.ModelOptions, omit(globalOptions, 'globalOptions'), target);
+  }
 }