+++ test/tests/shouldAdd.test.ts
@@ -7,7 +7,7 @@ import { Genders } from '../enums/genders';
 import { Alias, model as AliasModel } from '../models/alias';
 import { GetClassTestParent, GetClassTestParentModel, GetClassTestSub } from '../models/getClass';
 import { GetSet, GetSetModel } from '../models/getSet';
-import { model as InternetUser } from '../models/internetUser';
+import { InternetUserModel } from '../models/internetUser';
 import { BeverageModel as Beverage, InventoryModel as Inventory, ScooterModel as Scooter } from '../models/inventory';
 import { OptionsClass, OptionsModel } from '../models/options';
 import { model as User } from '../models/user';
@@ -94,7 +94,7 @@ export function suite() {
   });
 
   it(`should add dynamic fields using map`, async () => {
-    const user = await InternetUser.create({
+    const user = await InternetUserModel.create({
       socialNetworks: {
         twitter: 'twitter account',
         facebook: 'facebook account'
@@ -119,6 +119,9 @@ export function suite() {
     expect(user.sideNotes.get('day1')).to.have.property('content', 'day1');
     expect(user.sideNotes.get('day1')).to.have.property('link', 'url');
     expect(user.sideNotes.has('day2')).to.be.equal(true);
+
+    expect(user.sideNotes.get('day1')).to.not.have.property('_id');
+    expect(user.sideNotes.get('day2')).to.not.have.property('_id');
   });
 
   it('Should support dynamic references via refPath', async () => {